+++ basicalter/slice_test.go
@@ -37,14 +37,14 @@ func TestDelStringInSlice(t *testing.T) {
 	}
 }
 
-func TestReverseStrings(t *testing.T) {
+func TestReverseSlice(t *testing.T) {
 	sliceOfString := []string{"foo", "baz", "bar", "World", "Hello"}
 
-	newSlice := basicalter.ReverseStrings(sliceOfString)
+	basicalter.ReverseStrings(sliceOfString)
 
 	desiredSlice := []string{"Hello", "World", "bar", "baz", "foo"}
-	if !basiccheck.EqualStringSlice(newSlice, desiredSlice) {
-		t.Errorf("ReverseStrings fail to reverse slice: %v expected %v", newSlice, desiredSlice)
+	if !basiccheck.EqualStringSlice(sliceOfString, desiredSlice) {
+		t.Errorf("ReverseStrings fail to reverse slice: %v expected %v", sliceOfString, desiredSlice)
 	}
 }